WIP: trying to fix volume issues
don't bother installing chef infra client, the server has a version of
infra client inside it.

also try to deal with the madness that chef makes by putting tons of
stuff right inside /opt/opscode/embedded. This makes it impossbile to
start up an existing server keeping the configuration you already had
when it's not in the /var/opt/opscode volume -_-